@@ -1,31 +1,15 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
-import VueRouter from 'vue-router'
+import router from './router'
 import Loader from './components/Loader.vue'
-import ResultList from './components/ResultList.vue'
+
 
 import '../node_modules/bulma/css/bulma.css';
 import './assets/main.css'
 
-Vue.component('loader', Loader)
-Vue.use(VueRouter)
+const app = createApp(App)
 
-const routes = [{
-    path: '/',
-    component: App
-}, {
-    path: '/result',
-    component: ResultList
-}];
+app.use(router)
+app.component('Loader', Loader)
 
-const router = new VueRouter({
-    routes: routes
-});
-
-new Vue({
-    el: '#app',
-    router,
-    data: {
-        currentRoute: window.location.pathname
-    },
-});
+app.mount('#app')
